Move the survey component to its own file and handle the survey logging on its own file.
